feat: added agent control parameters
Added eyeOnTargetTime parameter, for full control on the amount of time the agent must keep eyesight on the target to complete each episode.
Added 'on target' related parameters to control when and for how long the agent keeps eyesight on the target.
Added m_TargetMeshRenderer to store the targets mesh renderer component on start.
Added set and reset on agent initialization and  on episode start.