feat: implemented agent rewards
Added agent rewards when the agent focuses on a target, and unrewards when lossing focus on target.
Start timing when agent has both eyes on the target, reset timer when lossing focus with either eye.
Unrewards agent when both eyes are not focused on the target.
Added rewards and episode end when agent manages to focus the target with both eyes for eyeOnTargetTime seconds.