feat: corrected viewport observations
Added peudo boolean flag for when the target isn't on the viewport.
Normalized out of bounds observations. The intention of the flag is to give the NN a way to ignore the out of bounds observations when the target isn't on the screen.
Corrected default VectorObservations to include the boolean flags